added interface for emailService
