+++ src/main/java/ma/nttdata/externals/commons/services/impl/EmailServiceImpl.java
@@ -1,0 +1,40 @@
+package ma.nttdata.externals.commons.services.impl;
+
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
+import lombok.RequiredArgsConstructor;
+import ma.nttdata.externals.commons.exception.InternalServerException;
+import ma.nttdata.externals.commons.services.EmailService;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class EmailServiceImpl implements EmailService {
+
+    private final JavaMailSender mailSender;
+
+    @Value("${app.mail.from}")
+    private String sender;
+
+    public void sendEmail(String to, String subject, String dynamicHtmlBody) {
+        if (mailSender == null) {
+            return;
+        }
+        try {
+            MimeMessage message = mailSender.createMimeMessage();
+            MimeMessageHelper helper = new MimeMessageHelper(message, true, "UTF-8");
+
+            helper.setFrom(sender);
+            helper.setTo(to);
+            helper.setSubject(subject);
+            helper.setText(dynamicHtmlBody, true);
+
+            mailSender.send(message);
+        } catch (MessagingException e) {
+            throw new InternalServerException("Failed to send email", e);
+        }
+    }
+}
