+++ src/main/java/ma/nttdata/externals/commons/services/impl/MailConfig.java
@@ -1,10 +1,9 @@
-package ma.nttdata.externals.commons.services;
+package ma.nttdata.externals.commons.services.impl;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
